+++ python/tests/t_envoy_logs.py
@@ -6,7 +6,7 @@ from kat.utils import ShellCommand
 from abstract_tests import AmbassadorTest, ServiceType, HTTP
 
 access_log_entry_regex = re.compile('^MY_REQUEST 200 .*')
-
+access_log_json_entry_regex = re.compile('.*"response_code":.*"200".*')
 
 class EnvoyLogTest(AmbassadorTest):
     target: ServiceType
@@ -39,3 +39,37 @@ config:
 
         for line in cmd.stdout.splitlines():
             assert access_log_entry_regex.match(line), f"{line} does not match {access_log_entry_regex}"
+
+class EnvoyJsonLogTest(AmbassadorTest):
+    target: ServiceType
+    log_path: str
+    log_format: str
+
+    def init(self):
+        if EDGE_STACK:
+            self.xfail = "Not yet supported in Edge Stack"
+
+        self.target = HTTP()
+        self.log_path = '/tmp/ambassador/ambassador_log.jsonl'
+        self.log_format = '{"response_code":"%RESPONSE_CODE%","authority":"%REQ(:AUTHORITY)%","user_agent":"%REQ(USER-AGENT)%","request_id":"%REQ(X-REQUEST-ID)%","upstream_host":"%UPSTREAM_HOST%"}'
+
+    def config(self):
+        yield self, self.format("""
+---
+apiVersion: ambassador/v1
+kind: Module
+name: ambassador
+ambassador_id: {self.ambassador_id}
+config:
+  envoy_log_path: {self.log_path}
+  envoy_log_type: json
+  envoy_log_format: '{self.log_format}'
+""")
+
+    def check(self):
+        cmd = ShellCommand("kubectl", "exec", self.path.k8s, "cat", self.log_path)
+        if not cmd.check("check envoy access log"):
+            pytest.exit("envoy access log does not exist")
+
+        for line in cmd.stdout.splitlines():
+            assert access_log_json_entry_regex.match(line), f"{line} does not match {access_log_json_entry_regex}"